TRDT-706 Remove duplicate hash utiliting and split into own spec
