+++ test/controllers/StubNonRepudiationServiceControllerSpec.scala
@@ -25,19 +25,14 @@ import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import utils.JsonUtils.jsonFromFile
 
-import java.nio.charset.StandardCharsets
-import java.security.MessageDigest
-import java.util.{Base64, UUID}
-import javax.inject.{Inject, Singleton}
 import scala.concurrent.Future
 import scala.util.matching.Regex
 
 class StubNonRepudiationServiceControllerSpec extends SpecBase with HeaderValidator {
 
-  val controller: NrsController = app.injector.instanceOf[NrsController]
+  private val controller: NrsController = app.injector.instanceOf[NrsController]
 
-  val v4UuidRegex: Regex = "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[4][0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$".r
-  val uuidRet: String = new UUID(0, 1).toString
+  private val v4UuidRegex: Regex = "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[4][0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$".r
 
   private def onSubmit(value: JsValue, withValidHeaders: Boolean = true): Future[Result] = {
     val request: FakeRequest[JsValue] = if (withValidHeaders) {
@@ -51,43 +46,7 @@ class StubNonRepudiationServiceControllerSpec extends SpecBase with HeaderValida
     controller.submit().apply(request)
   }
 
-
-  @Singleton
-  class HashUtil @Inject()() {
-
-    private def sha256: MessageDigest = MessageDigest.getInstance("SHA-256")
-
-    def encode(value: String): String = Base64.getUrlEncoder.encodeToString(value.getBytes(StandardCharsets.UTF_8))
-
-    //TODO ensure that service is using getUrlEncoder.
-    def getHash(value: String): String = sha256.digest(value.getBytes()).map("%02x" format _).mkString
-
-  }
-
   "StubNonRepudiationServiceController onSubmit" should {
-
-    // Use this to quickly generate the payload data and checksum.
-//            "Generate output" in {
-//              val hashUtil: HashUtil = new HashUtil
-//              val payload = "{\"reportId\":\"a365c0b4-06e3-4fef-a555-16fd0877dc7c\"}"
-//              val payloadBase64 = hashUtil.encode(payload)
-//              val payloadSha = hashUtil.getHash(payload)
-//              println (s"payload original::${payload}")
-//              println (s"payloadBase64::${payloadBase64}")
-//              println (s"payloadSha::${payloadSha}")
-//
-//              true must be( true )
-//            }
-
-    "generate sha256 and base64 values for payload" in {
-      val hashUtil: HashUtil = new HashUtil
-      val payload = "{\"reportId\":\"a365c0b4-06e3-4fef-a555-16fd0877dc7c\"}"
-      val sha = hashUtil.getHash(payload)
-      val base64 = hashUtil.encode(payload)
-
-      sha must be("bb895fc5f392e75750784dc4cc3fe9d4055516dfe012c3ae3dc09764dfa19413")
-      base64 must be("eyJyZXBvcnRJZCI6ImEzNjVjMGI0LTA2ZTMtNGZlZi1hNTU1LTE2ZmQwODc3ZGM3YyJ9")
-    }
 
     "check message payload(encode) payloadSha256Checksum(encode) can be read from file" in {
       val json = jsonFromFile("/validNrsEventAcknowledgeChecksumSha.json")
@@ -102,7 +61,7 @@ class StubNonRepudiationServiceControllerSpec extends SpecBase with HeaderValida
       val result = onSubmit(json)
 
       status(result) must be(ACCEPTED)
-      (contentAsJson(result) \ "nrSubmissionId").as[String] must be(uuidRet)
+      (contentAsJson(result) \ "nrSubmissionId").as[String] must fullyMatch regex v4UuidRegex
     }
 
     "return 401 Unauthorised when invalid headers received" in {
